Create a multithreaded post-processing stage
